% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy_loader.R
\name{load_cpp_strategies}
\alias{load_cpp_strategies}
\title{Load C++ strategies from a specified directory}
\usage{
load_cpp_strategies(directory = "strategies", env = .GlobalEnv)
}
\arguments{
\item{directory}{Character string. The path to the folder containing .cpp files.}

\item{env}{An environment where the functions will be loaded (default is .GlobalEnv).}
}
\value{
A character vector containing the names of the strategies successfully compiled.
}
\description{
This function compiles and sources all .cpp files in a directory.
Note that for consistency, it is assumed that the function name exported
in the C++ code matches the filename (e.g., 'my_strat.cpp' should
export a function named 'my_strat').
}
\note{
This function does not include a \code{prefix} argument, unlike the Python
equivalent. This is because C++ functions are exported with fixed symbol
names during the compilation process via \verb{// [[Rcpp::export]]}. While we
can capture the filenames for tracking, we cannot programmatically rename
the underlying compiled symbols in the same way we can with Python objects.
}
