% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy_loader.R
\name{load_python_strategies}
\alias{load_python_strategies}
\title{Load Python strategies from a specified directory}
\usage{
load_python_strategies(
  directory = "strategies",
  prefix = "py_",
  env = .GlobalEnv
)
}
\arguments{
\item{directory}{Character string. The path to the folder containing .py files.}

\item{prefix}{Character string. A prefix to prepend to the loaded functions (e.g., "py_").}

\item{env}{An environment where the functions will be assigned (default is .GlobalEnv).}
}
\value{
A character vector containing the names of the strategies successfully loaded.
}
\description{
Load Python strategies from a specified directory
}
