% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payoff.R
\name{print.flkbrg_payoff}
\alias{print.flkbrg_payoff}
\title{Print method for flkbrg_payoff}
\usage{
\method{print}{flkbrg_payoff}(x, ...)
}
\arguments{
\item{x}{The flkbrg_payoff object.}

\item{...}{Additonal Arguments. Currently not used.}
}
\description{
This is a generic print function to beautify the printing of flkbrg_payoff
objects on the console. It is not dynamic and does not react to anything.
There is no love in this function. Just a simple helper to not mess up stdout.
}
