% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tournament.R
\name{print.flkbrg_tournament}
\alias{print.flkbrg_tournament}
\title{Print Method for flkbrg_tournament Objects}
\usage{
\method{print}{flkbrg_tournament}(x, ...)
}
\arguments{
\item{x}{An object of class \code{flkbrg_tournament}.}

\item{...}{Further arguments passed to or from other methods (ignored).}
}
\value{
\code{x}, invisibly.
}
\description{
Prints a concise summary of a tournament result, showing the standings
table and listing the names of the additional returned items.
}
