% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategies.R
\name{r_traitor}
\alias{r_traitor}
\title{Traitor strategy.}
\usage{
r_traitor(my_hist, opp_hist, game_info = NA)
}
\arguments{
\item{my_hist}{A character vector of the player's own previous moves.}

\item{opp_hist}{A character vector of the opponent's previous moves.}

\item{game_info}{A list containing metadata about the current game state.}
}
\description{
Traitor strategy.
}
