% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{match}
\alias{match}
\title{Simulate a single match between two players using a flkbrg_payoff object.}
\usage{
match(
  strategy1,
  strategy2,
  n_rounds = 200,
  include_info = TRUE,
  payoff = flkbrg_payoff()
)
}
\arguments{
\item{strategy1}{A function representing the first player's decision logic.}

\item{strategy2}{A function representing the second player's decision logic.}

\item{n_rounds}{Integer specifying the number of iterations per match.
Defaults to 200, which is the original value from Axelrods
first tournament.}

\item{include_info}{Logical indicating whether metadata should be passed to strategies.}

\item{payoff}{An object of class \code{flkbrg_payoff}. Use flkbrg_payoff() to create one.}
}
\value{
A list containing names, scores, and a history data frame.
}
\description{
Simulate a single match between two players using a flkbrg_payoff object.
}
